Debugger: Tilemap viewer - Minor visual tweaks to mode 7 overlay

+++ NewUI/Debugger/Controls/PictureViewer.cs
@@ -492,7 +492,7 @@ namespace Mesen.Debugger.Controls
 		public Point Start;
 		public Point End;
 		public Color Color;
-		public int? Width;
+		public double? Width;
 		public IDashStyle? DashStyle;
 	}
 }
